requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
docstring_parser==0.16
executing==2.0.1
Flask==3.0.2
Flask-Cors==4.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.45.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-generativeai==0.4.1
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
ipykernel==6.29.4
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
langchain-core==0.1.36
langchain-google-genai==1.0.1
langchain-google-vertexai==0.1.2
langsmith==0.1.38
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
orjson==3.10.0
packaging==23.2
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
rsa==4.9
shapely==2.0.3
six==1.16.0
stack-data==0.6.3
tenacity==8.2.3
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
types-protobuf==4.24.0.20240311
types-requests==2.31.0.20240311
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1

# Optional dependencies for horizontal scaling (production)
# Install these for distributed deployment across multiple instances:
redis==5.0.1  # For distributed sessions and caching
prometheus-client==0.20.0  # For Prometheus metrics export
sentry-sdk[flask]==2.0.0  # For error tracking and performance monitoring

# ✅ NEW: Security dependencies
PyJWT==2.8.0  # JWT authentication
cryptography==42.0.5  # Encryption for secrets management
bcrypt==4.1.2  # Password hashing
SQLAlchemy==2.0.28  # ORM for database operations

# ✅ NEW: GDPR Compliance & Data Retention
APScheduler==3.10.4  # Scheduled data retention cleanup

# ✅ NEW: Testing dependencies (Phase 6)
pytest==8.0.2  # Testing framework
pytest-cov==4.1.0  # Coverage reporting
pytest-mock==3.12.0  # Mocking utilities
pytest-asyncio==0.23.5  # Async test support
pytest-timeout==2.2.0  # Test timeout handling
pytest-xdist==3.5.0  # Parallel test execution
hypothesis==6.98.0  # Property-based testing
faker==24.0.0  # Fake data generation for tests
responses==0.25.0  # Mock HTTP responses
freezegun==1.4.0  # Mock datetime for tests

# ✅ NEW: Security testing tools (Phase 6.7)
bandit==1.7.6  # Python security scanner
safety==3.0.1  # Dependency vulnerability scanner
pip-audit==2.7.0  # Alternative dependency scanner
detect-secrets==1.4.0  # Secret detection
